refactor: 가시성 변경 및 클래스 이동
