+++ src/main/java/study/Calculator.java
@@ -1,10 +1,8 @@
 package study;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
-import java.util.function.BiFunction;
 import java.util.regex.Pattern;
 
 class Calculator {
@@ -37,7 +35,7 @@ class Calculator {
         return result;
     }
 
-    static class SubFormula {
+    private static class SubFormula {
         public static SubFormula create(String formula) {
             try {
                 String leftHandSide = extractLeftHandSide(formula);
@@ -104,34 +102,4 @@ class Calculator {
         }
     }
 
-    enum Operator {
-        plus("+", (left, right) -> left + right), //
-        minus("-", (left, right) -> left - right), //
-        multiply("*", (left, right) -> left * right), //
-        divide("/", (left, right) -> {
-            if (left % right != 0) {
-                throw new IllegalArgumentException("나눗셈은 정수로 떨어져야 합니다.");
-            }
-            return left / right;
-        });
-
-        private final String sign;
-        private final BiFunction<Long, Long, Long> operationImpl;
-
-        Operator(String sign, BiFunction<Long, Long, Long> operationImpl) {
-            this.sign = sign;
-            this.operationImpl = operationImpl;
-        }
-
-        public static Operator valueOfSign(String sign) {
-            return Arrays.stream(values()) //
-                    .filter(value -> value.sign.equals(sign)) //
-                    .findAny() //
-                    .orElseThrow(() -> new IllegalArgumentException("올바른 연산자가 아닙니다."));
-        }
-
-        public Long apply(Long left, Long right) {
-            return operationImpl.apply(left, right);
-        }
-    }
 }
